Files for the chat system
These are all the files required to make a FaceBook style chat system. All that needs to be done now is to create buttons on the active enquiries page that allow for the chat log to be opened. I will need to create a new method when that happens which removes then adds the event listeners, this can only be done when the enquiries system is complete.
The files do the following
 -chat.php-this is the class which contains the relevant methods like sendMessage
 -chatManager-this is the javascript file which is placed in any page you wish a chat to exist in
 -refreshMessages.php and sendMessages.php-these are called by the javascript file when the relevant action is performed